Add gitea Ansible role to list (#117)
For your consideration, this PR adds an Ansible role that I've cobbled together to run two Gitea instances on the same VPS (but the logic in the role should allow an arbitrary number of instances).

It is something of a hobby project of mine, but I'd be pleased to receive any comments or suggestions on how to improve the role.

Co-authored-by: taha@asks2 <taha@chepec.se>
Reviewed-on: https://gitea.com/gitea/awesome-gitea/pulls/117
Reviewed-by: John Olheiser <john+gitea@jolheiser.com>
Reviewed-by: delvh <dev.lh@web.de>
Co-authored-by: solarchemist <gitea@solarchemist.se>
Co-committed-by: solarchemist <gitea@solarchemist.se>